+++ src/pages/Contact/Contact.tsx
@@ -1,9 +1,9 @@
 import { Container, Row } from "react-bootstrap";
-import Footer from "../../components/Footer";
-import ContactImage from "./image";
-import Cards from "./cards";
-import ContactForm from "./form";
-import Map from "./map";
+import Footer from "@/components/Footer";
+import ContactImage from "./Image";
+import Cards from "./Cards";
+import ContactForm from "./Form";
+import Map from "./Map";
 
 const ContactPage = () => {
   return (