+++ src/pages/home/section.tsx
@@ -12,10 +12,10 @@ const Section = () => {
         <div className="row justify-content-center">
           <div className="col-md-8 text-center">
             <h4 className="display-6 text-white mb-4">{t("about")}</h4>
-            <p className="text-white text-justify fs-5">
+            <p className="text-white text-justify fs-6">
               {t("about_text.first")}
             </p>
-            <p className="text-white text-justify mt-3 fs-5">
+            <p className="text-white text-justify mt-3 fs-6">
               {t("about_text.second")}
             </p>
           </div>