+++ vite.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import path from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -18,5 +19,11 @@ export default defineConfig({
   },
   server: {
     port: 3000,
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src'), // Adds `@` as an alias for the `src` directory
+      // Add more aliases here if needed
+    },
   },
 });